Get rid of special case for CRLF
The existing code treated CR and LF as special cases of the Control
grapheme category, for reasons that weren't very good. This patch gets
rid of that and just handles GB3 in the pair lookup. That should
improve performance in the rope case, as it will cut down on the
amount of pre-context requested when a chunk begins with LF.