tests: ensure version numbers in README are always up to date
This adds a test that will check the examples in README.md for the
current package version.

It also checks that there is a Markdown heading mentioning (at least)
the current version. This serves as a simple check that the change log
section is updated when the version number is bumped.

Related to #30, #41, #47, and #48.