Optimization for grapheme iteration.
Typical text involves a lot of chars one-after-another that
are in a similar scalar value range and in the same category.  This
commit takes advantage of that to speed up grapheme iteration by
caching the range and category of the last char, and always checking
against that cache first before doing a binary search on the whole
grapheme category table.

This results in significant speed ups on many texts, in some cases
up to 2x faster.