Link to creating methods in struct documentation
The Rust standard library has a nice practice for documentation for public
structs. Documentation for these structs include links to the method that
creates them. This is very useful for beginners who may stumble upon the
documentation for the struct when they would really like documentation for the
method.

Add documentation for all of the structs that are created by methods on the
`UnicodeSegmentation` trait with links back to the creating methods.