Implement Debug on all public structs
Some types in this crate don't have Debug implemented, but it's annoying
to use a non-Debug type in a struct with Debug derived. I just used the
automatic impl from `#[derive(Debug)]` for all of these (although in the
future it might be good to have custom implementations for types with
complex internals).

Closes #95.